+++ package.json
@@ -23,6 +23,7 @@
     "@types/node": "17.0.30",
     "@types/react": "18.0.8",
     "@types/react-dom": "18.0.3",
+    "@types/react-intl": "^3.0.0",
     "eslint": "8.14.0",
     "eslint-config-next": "12.1.5",
     "husky": "7.0.4",
